+++ src/private/cli/GetAwesomeHostUi.ps1
@@ -1,0 +1,4 @@
+function Get-AwesomeHostUi {
+    return $Host.UI
+}
+