+++ src/private/cli/ReadAwesomeHost.ps1
@@ -1,3 +1,4 @@
+
 function Read-AwesomeHost {
     [CmdletBinding()]
     param (
@@ -5,28 +6,11 @@ function Read-AwesomeHost {
         [pscustomobject]
         $Ask
     )
-    ## For standard questions
-    if ($null -eq $Ask.Choice) {
-        do {
-            $response = $Host.UI.Prompt($Ask.Caption, $Ask.Message, $Ask.Prompt)
-        } while ($Ask.Default -eq 'MANDATORY' -and [string]::IsNullOrEmpty($response.Values))
 
-        if ([string]::IsNullOrEmpty($response.Values)) {
-            $result = $Ask.Default
-        } else {
-            $result = $response.Values
-        }
+    $hostUi = Get-AwesomeHostUi
+    if ($null -eq $Ask.Choice) {
+        return Read-AwesomeStandardPrompt -Ask $Ask -HostUi $hostUi
     }
-    ## For Choice based
-    if ($Ask.Choice) {
-        $Cs = @()
-        $Ask.Choice.Keys | ForEach-Object {
-            $Cs += New-Object System.Management.Automation.Host.ChoiceDescription "&$_", $($Ask.Choice.$_)
-        }
-        $options = [System.Management.Automation.Host.ChoiceDescription[]]($Cs)
-        $IndexOfDefault = $Cs.Label.IndexOf('&' + $Ask.Default)
-        $response = $Host.UI.PromptForChoice($Ask.Caption, $Ask.Message, $options, $IndexOfDefault)
-        $result = $Cs.Label[$response] -replace '&'
-    }
-    return $result
+
+    return Read-AwesomeChoicePrompt -Ask $Ask -HostUi $hostUi
 }