+++ src/private/cli/TestAwesomePromptRequiresRetry.ps1
@@ -1,0 +1,9 @@
+function Test-AwesomePromptRequiresRetry {
+    param(
+        [Parameter(Mandatory)][pscustomobject]$Ask,
+        [Parameter(Mandatory)][object]$Response
+    )
+
+    return $Ask.Default -eq 'MANDATORY' -and [string]::IsNullOrEmpty($Response.Values)
+}
+